relocate httpProfile into util package
