+++ util/httpProfile_test.go
@@ -14,4 +14,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package collector
+package util
